refactor(attributes): move enums to core
