+++ packages/opentelemetry-core/src/index.ts
@@ -18,6 +18,7 @@ export * from './common/ConsoleLogger';
 export * from './common/NoopLogger';
 export * from './common/time';
 export * from './common/types';
+export * from './common/attributes';
 export * from './ExportResult';
 export * from './version';
 export * from './context/context';
